Egen komponent for AvkortingInntektForm
